Fix some 64-bit warnings in VS2012.

+++ trunk/cgreen/src/message_formatting.c
@@ -23,7 +23,7 @@ static int find_index_of_difference(char *expected, char *actual, size_t size_to
 
     while (size_to_compare--) {
         if (*expectedp++ != *actualp++) {
-            return actualp - actual;
+            return (int)(actualp - actual);
         }
     }
